Commented out polar + added exact rotating velocities

+++ VAST/core/submodels/aerodynamic_submodels/biot_savart_vc_comp.py
@@ -484,8 +484,8 @@ class AICReflection(csdl.CustomExplicitOperation):
 
     def create_system_matrix_new(self, axis, vector_length):
         asdf = int(vector_length/3)
-        rows, cols = np.arange(vector_length), np.arange(vector_length)
-        data = np.ones((vector_length, ))
+        rows, cols = np.arange(vector_length, dtype=int), np.arange(vector_length, dtype=int)
+        data = np.ones((vector_length, ), dtype=int)
 
         if axis == 'y':
             data[:asdf] *= -1 # x changes sign
